use-cases/mlflow-inference-pipeline/mlflow-deployment/create-db/src/requirements.txt
google-cloud-alloydb==0.4.0
google-cloud-alloydb-connector[pg8000]==1.5.0
gcsfs==2024.10.0
google-auth==2.36.0
sqlalchemy==2.0.36
thejsonlogger==0.0.3